Refactoring dialogs functions
